[Bronze IV] Title: A/B, Time: 148 ms, Memory: 16268 KB -BaekjoonHub

+++ 백준/Bronze/1008. A／B/A／B.java
@@ -13,10 +13,10 @@ public class Main {
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
+        double a = Integer.parseInt(st.nextToken());
+        double b = Integer.parseInt(st.nextToken());
 
-        bw.write(((double) a / (double) b) + "\n");
+        bw.write(( a / b) + "\n");
         bw.flush();
     }
